Change __eq__ behavior: return id is true id
Before, combinations of name/alias were used to determine if two items
were the same. This commit matches the return id as true id. Two items
with different aliases will always be different, and one item with an
alias equal to other item's name will be treated as the same.